standarize size type (#747)
* serial_size_t -> size_t

* fix static_cast<size_t>

* apply clang-format

* fix deserialization

* rename routines

* add beru's patch

* add move semantics

* fix warnings

* clang-format

* fix deconvolutional tests

* clang-format

* fix test_tensor

* Update test code

* fix master

* Fix compilation errors

* reduce warnings

* Adjust serialization implementation to pass tests

* clang-format

+++ test/test_quantization.h
@@ -94,8 +94,8 @@ TEST(quantization_utils, quantized_to_float) {
 
 TEST(quantization_utils, avoid_bias) {
   for (int i = 0; i < 256; ++i) {
-    const float as_float =
-      core::kernels::quantized_to_float<uint8_t>(i, 0.0f, 2.0f);
+    const float as_float = core::kernels::quantized_to_float<uint8_t>(
+      static_cast<uint8_t>(i), 0.0f, 2.0f);
     const int back_to_int =
       core::kernels::float_to_quantized<uint8_t>(as_float, 0.0f, 2.0f);
     EXPECT_EQ(i, back_to_int);
